requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
openai==1.58.1
jinja2==3.1.4
python-multipart==0.0.18
pytest==8.3.4
httpx==0.28.1
pytest-playwright==0.7.1
slowapi==0.1.9
